backend/requirements.txt
fastapi
uvicorn
openai
python-dotenv  # Optional, for API keys
